+++ src/main/java/ensino/configuracoes/view/panels/curso/CursoFieldsPanel.java
@@ -122,7 +122,7 @@ public class CursoFieldsPanel extends DefaultFieldsPanel {
         return map;
     }
 
-    private void setFieldValues(Integer codigo, String nome,
+    private void setFieldValues(Long codigo, String nome,
             Campus iCampus, NivelEnsino nivelEnsino) {
         txtId.setText(codigo.toString());
         txtNome.setText(nome);
@@ -135,7 +135,7 @@ public class CursoFieldsPanel extends DefaultFieldsPanel {
 
     @Override
     public void setFieldValues(HashMap<String, Object> mapValues) {
-        Integer codigo = (Integer) mapValues.get("id");
+        Long codigo = (Long) mapValues.get("id");
         
         setFieldValues(codigo, (String) mapValues.get("nome"),
                 (Campus) mapValues.get("campus"),
