+++ src/main/java/ensino/configuracoes/view/panels/unidadeCurricular/UnidadeCurricularFields.java
@@ -153,7 +153,7 @@ public class UnidadeCurricularFields extends DefaultFieldsPanel {
         return map;
     }
 
-    private void setFieldValues(Integer id, String nome, Integer nAulasTeoricas,
+    private void setFieldValues(Long id, String nome, Integer nAulasTeoricas,
             Integer nAulasPraticas, Integer cargaHoraria,
             String ementa, Curso curso) {
         txtId.setText(id.toString());
@@ -168,7 +168,7 @@ public class UnidadeCurricularFields extends DefaultFieldsPanel {
     @Override
     public void setFieldValues(HashMap<String, Object> mapValues) {
         setFieldValues(
-                (Integer) mapValues.get("id"),
+                (Long) mapValues.get("id"),
                 (String) mapValues.get("nome"),
                 (Integer) mapValues.get("nAulasTeoricas"),
                 (Integer) mapValues.get("nAulasPraticas"),
