+++ src/main/java/ensino/planejamento/view/panels/planoDeEnsino/DetalhamentoFieldsPanel.java
@@ -73,7 +73,7 @@ import javax.swing.event.ChangeListener;
  */
 public class DetalhamentoFieldsPanel extends DefaultFieldsPanel {
 
-    private Integer sequencia;
+    private Long sequencia;
     private SemanaLetiva semanaLetiva;
     private GenJLabel lblSemana;
 
@@ -506,7 +506,7 @@ public class DetalhamentoFieldsPanel extends DefaultFieldsPanel {
 
                     BaseObject selected = (BaseObject) comboMetodo.getSelectedItem();
                     if (selected != null && tipoMetodo != null) {
-                        Integer id = 1;
+                        Long id = 1L;
                         if (!metodologiaTableModel.isEmpty()) {
                             Metodologia mTemp = metodologiaTableModel.getMax(Comparator.comparing(a -> a.getId().getSequencia()));
                             id = mTemp.getId().getSequencia() + 1;
